requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.9.11
cffi==1.14.1
chardet==3.0.4
Click==7.0
cloud-init==19.2
cloudpickle==1.4.1
cmake==3.15.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
dask==2.19.0
decorator==4.4.2
distro-info===0.18ubuntu0.18.04.1
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
firebase-admin==4.3.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
MarkupSafe==1.1.1
matplotlib==3.2.2
msgpack==1.0.0
netifaces==0.10.4
networkx==2.4
numpy==1.17.2
oauthlib==2.0.6
opencv-python==4.1.1.26
PAM==0.4.2
Pillow==6.2.0
protobuf==3.12.2
psycopg2==2.8.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-dotenv==0.12.0
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
requests==2.22.0
requests-unixsocket==0.1.5
rsa==4.6
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.5.0
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
ssh-import-id==5.7
systemd-python==234
toolz==0.10.0
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.6
Werkzeug==0.16.0
zope.interface==4.3.2
